+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- (cli) Added a `plt switch` subcommand which is the equivalent of running `plt clone --force` and then running `plt cache-repo` and then running `plt apply`. This allows a common task (switching the version of a pallet and applying it immediately) to be run with a single command, for a simpler user experience.
+
 ## 0.5.0 - 2024-01-14
 
 ### Added